requirements.txt
django
psycopg2-binary
psycopg2
pillow
djangorestframework
django-filter
requests
stripe
celery
eventlet
django-celery-beat
python-dotenv
attrs
coreapi
tasks
redis
flake8
djangorestframework_simplejwt
drf_yasg
django-cors-headers
gunicorn
